refactored salaryrange to classification
